+++ src/main/java/com/gamemoonchul/application/converter/PostConverter.java
@@ -1,5 +1,6 @@
 package com.gamemoonchul.application.converter;
 
+import com.gamemoonchul.application.member.MemberConverter;
 import com.gamemoonchul.domain.entity.Member;
 import com.gamemoonchul.domain.entity.Post;
 import com.gamemoonchul.infrastructure.web.dto.PostResponseDto;
