+++ src/main/java/com/gamemoonchul/application/member/MemberConverter.java
@@ -1,4 +1,4 @@
-package com.gamemoonchul.application.converter;
+package com.gamemoonchul.application.member;
 
 import com.gamemoonchul.config.apple.AppleCredential;
 import com.gamemoonchul.config.oauth.user.OAuth2Provider;
